add hook for thread actions
